Reorganize tests and test data.
Apps which have many tests should store them in proper Python modules along with the data they rely on.
