+++ backend/tests/data/ftccsdartic:oai:hal.archives-ouvertes.fr:hal-00939473.xml
