+++ backend/tests/test_oadoi.py
@@ -2,13 +2,19 @@
 
 
 import pytest
+import os
 import django.test
 from backend.oadoi import OadoiAPI
 from papers.models import Paper
 
 @pytest.mark.usefixtures("load_test_data")
 class OadoiAPITest(django.test.TestCase):
-    
+
+    @classmethod
+    def setUpClass(cls):
+        super(OadoiAPITest, cls).setUpClass()
+        cls.testdir = os.path.dirname(os.path.abspath(__file__))
+
     def test_ingest_dump(self):
         doi = '10.1080/21645515.2017.1330236'
         p = Paper.create_by_doi(doi)
@@ -17,7 +23,7 @@ class OadoiAPITest(django.test.TestCase):
 
         # then load an OAdoi dump
         oadoi = OadoiAPI()
-        oadoi.load_dump('devutils/sample_unpaywall_snapshot.jsonl.gz')
+        oadoi.load_dump(os.path.join(self.testdir, 'data/sample_unpaywall_snapshot.jsonl.gz'))
 
         # the paper is now OA, yay!
         p = Paper.get_by_doi(doi)
