+++ papers/tests/data/citeseerx_record_without_date.xml
