+++ papers/tests/data/hal_record.xml
