+++ publishers/tests.py
@@ -19,7 +19,7 @@
 
 from backend.romeo import fetch_journal
 from papers.models import Paper
-from papers.testpages import RenderingTest
+from papers.tests.test_pages import RenderingTest
 
 
 class JournalPageTest(RenderingTest):
