+++ statistics/conftest.py
@@ -3,7 +3,7 @@ from django.core.management import call_command
 
 from papers.models import Institution
 from papers.models import Department
-from backend.tests import get_researcher_by_name
+from backend.tests.test_generic import get_researcher_by_name
 from backend.orcid import OrcidPaperSource
 
 @pytest.fixture(scope='class')
