+++ upload/tests/data/red-circle.png
